pyproject.toml
[project]
name = "cdpdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.8",
    "playwright>=1.58.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.12.5",
    "pytz>=2025.2",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.0",
]
